backend/requirements.txt
appnope==0.1.3
asgiref==3.8.1
asttokens==2.0.5
backcall==0.2.0
decorator==5.1.1
Django==3.2.10
django-cors-headers==3.10.0
djangorestframework==3.12.4
executing==0.8.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pytz==2024.2
six==1.16.0
sqlparse==0.5.3
stack-data==0.2.0
traitlets==5.2.2.post1
typing_extensions==4.12.2
wcwidth==0.2.5
